Autonomous Vehicle Simulation with Artificial Intelligence
As part of the Erciyes University Graduation Project, an autonomous vehicle simulation was developed using artificial intelligence and artificial neural networks. In this study, the NEAT (Neural Evolution of Augmenting Topologies) algorithm was utilized to enable vehicles to process environmental data and make independent decisions. A two-dimensional simulation environment was created using the Python programming language and the Pygame library. The project aims to evaluate the integration potential of autonomous driving systems into real-world applications and contribute to improving their performance.